fix: read_header broke PDFStreamParser
